Remove specific test and fall back on paging
