+++ test/AutoRest.TestServer.Tests/paging.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
 using AutoRest.TestServer.Tests.Infrastructure;
 using Azure;
@@ -845,6 +846,23 @@ namespace AutoRest.TestServer.Tests
         public void PagingModelsAreHidden()
         {
             Assert.IsFalse(typeof(ProductResult).IsPublic);
+        }
+
+        [Test]
+        public void RestClientDroppedMethods()
+        {
+            var client = typeof(PagingRestClient);
+            Assert.IsNull(client.GetMethod ("CreateNextFragmentNextPageRequest", BindingFlags.Instance | BindingFlags.NonPublic), "CreateNextFragmentNextPageRequest should not be generated");
+            Assert.IsNull(client.GetMethod ("NextFragmentNextPageAsync", BindingFlags.Instance | BindingFlags.NonPublic), "NextFragmentNextPageAsync should not be generated");
+            Assert.IsNull(client.GetMethod ("NextFragmentNextPage", BindingFlags.Instance | BindingFlags.NonPublic), "NextFragmentNextPage should not be generated");
+        }
+
+        [Test]
+        public void ClientDroppedMethods()
+        {
+            var client = typeof(PagingClient);
+            Assert.IsNull(client.GetMethod ("NextFragmentAsync", BindingFlags.Instance | BindingFlags.NonPublic), "NextFragmentAsync should not be generated");
+            Assert.IsNull(client.GetMethod ("NextFragment", BindingFlags.Instance | BindingFlags.NonPublic), "NextFragment should not be generated");
         }
     }
 }
