--- test/TestProjects/PageNextLink/pagenextlink.json
@@ -1,131 +1,0 @@
-{
-    "swagger": "2.0",
-    "info": {
-      "title": "AutoRest Paging Test Service",
-      "description": "Long-running Operation for AutoRest",
-      "version": "1.0.0"
-    },
-    "host": "localhost:3000",
-    "schemes": [
-      "http"
-    ],
-    "produces": [
-      "application/json"
-    ],
-    "consumes": [
-      "application/json"
-    ],
-    "paths": {
-      "/paging/multiple/fragment/{tenant}": {
-        "get": {
-          "x-ms-pageable": { "nextLinkName": "odata.nextLink", "itemName": "values", "operationName": "Paging_nextFragment" },
-          "operationId": "Paging_getMultiplePagesFragmentNextLink",
-          "description": "A paging operation that doesn't return a full URL, just a fragment",
-          "parameters": [
-            {
-              "name": "api_version",
-              "in": "query",
-              "required": true,
-              "type": "string",
-              "description": "Sets the api version to use."
-            },
-            {
-              "name": "tenant",
-              "in": "path",
-              "required": true,
-              "type": "string",
-              "description": "Sets the tenant to use."
-            }
-          ],
-          "responses": {
-            "200": {
-              "description": "Paging response with a fragment nextLink",
-              "schema": {
-                "$ref": "#/definitions/OdataProductResult"
-              }
-            },
-            "default": {
-              "description": "Unexpected error"
-            }
-          }
-        }
-      }
-    },
-    "x-ms-paths": {
-        "/paging/multiple/fragment/{tenant}/{nextLink}": {
-          "get": {
-            "x-ms-pageable": { "nextLinkName": "odata.nextLink", "itemName": "values", "operationName": "Paging_nextFragment" },
-            "operationId": "Paging_nextFragment",
-            "description": "A paging operation that doesn't return a full URL, just a fragment",
-            "parameters": [
-              {
-                "name": "api_version",
-                "in": "query",
-                "required": true,
-                "type": "string",
-                "description": "Sets the api version to use."
-              },
-              {
-                "name": "tenant",
-                "in": "path",
-                "required": true,
-                "type": "string",
-                "description": "Sets the tenant to use."
-              },
-              {
-                "name": "nextLink",
-                "in": "path",
-                "required": true,
-                "type": "string",
-                "description": "Next link for list operation.",
-                "x-ms-skip-url-encoding": true
-              }
-            ],
-            "responses": {
-              "200": {
-                "description": "Paging response with a fragment nextLink",
-                "schema": {
-                  "$ref": "#/definitions/OdataProductResult"
-                }
-              },
-              "default": {
-                "description": "Unexpected error"
-              }
-            }
-          }
-        }
-    },
-    "definitions": {
-        "OdataProductResult": {
-            "type":  "object",
-            "properties": {
-                "values": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/Product"
-                    }
-                    },
-                    "odata.nextLink": {
-                    "type": "string"
-                }
-            }
-        },
-        "Product": {
-            "type":  "object",
-            "properties": {
-                "properties": {
-                    "type": "object",
-                    "properties": {
-                        "id": {
-                        "type": "integer"
-                        },
-                        "name": {
-                        "type": "string"
-                        }
-                    }
-                }
-            }
-        }
-      }
-  }
-
