--- test/TestProjects/PageNextLink/readme.md
@@ -1,8 +1,0 @@
-# PageNextLink
-### AutoRest Configuration
-> see https://aka.ms/autorest
-
-``` yaml
-require: $(this-folder)/../../../readme.md
-input-file: $(this-folder)/pagenextlink.json
-```
